mc_att_control: refactor to ModuleBase & add module documentation
